+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.6.0
+Release date: February, 8th 2021
+
+### Added
+PLGOPNS-379: Add support for OpenCart 3.0.3.7  
+PLGOPNS-373: Add company_name to CustomerDetails object in order request
+
+### Changed
+PLGOPNS-377: Replace getPaymentLink() with getPaymentUrl() to prepare future deprecation of the method in the PHP-SDK
+PLGOPNS-376: Improvements in callback notification function
+
+### Fixed
+PLGOPNS-375: Remove unneded code related with the shopping cart of the transaction    
+PLGOPNS-378: Fix bug for fixed coupons applied before taxes
+
+
 ## 3.5.0
 Release date: December, 18th 2020